requirements.txt
EbookLib
python-pptx
python-docx
trafilatura
pymupdf
google-api-core
pytube
openai-whisper